requirements.txt
esper
pygame-ce
pygbag
pyinstaller
